occupancy determination changed
occupancy is now determined independently and added as an input to this function. The check for occ/unocc having at least two levels is removed. When creating a model, the code now checks the number of levels present in ftow. If ftow has only one level, it is removed from the dframe.