Added shifting functionality to the aggregate function
TLDR; Added time shifting functionality to the aggregate function that handles shifting weather data in a more robust way than create_dataframe.

Added time shift functionality to the aggregate function as an optional argument. This is to address issues with time shifting that was handled in the create_dataframe function that attempted to shift all load and independent variable data forward to the end of the period to match normal weather data whose timestamps represent the end of the period. However, the handling of this timeshifting appears to have been done after aggregation, leading to larger than necessary shifts for hourly aggregation. The new shift functionality for the aggregate function shifts only the temp time series back by one increment of its interval prior to aggregation if the shift_normal_weather flag is TRUE. This eliminates the issue of everything getting shifted forward by an hour, day, or month when aggregating to those respective intervals which occurred in create_dataframe.