WIP: Aggregate function monthly data handling overhaul
Extensive re-write of how monthly data is handled in the aggregate function. All for loops have been removed to keep the style and syntax consistent with other parts of this function and hopefully increase readability. Support for additional independent variable aggregation has also been added to monthly aggregation. Further testing is required to confirm that the behavior is consistent with the create_dataframe function's handling of additional independent variables when aggregating to monthly data.