Update to Aggregate function to output of intervals used in monthly data aggregation
Updated the aggregate function to further clarify what date ranges are being used to aggregate data when monthly aggregation is selected. Now, monthly data aggregation will create a dataframe with 2 additional columns, period_start and period_end. The current assumption is that usage data provided on a monthly interval (or longer) has time stamps representing the beginning of the usage period. For example, if a time series has (5/15/23, 200 kWh) and (6/15/23, 170 kWh), then we assume the usage period associated with 200 kWh is 5/15/23-6/15/23 (not inclusive of the final date). Because the nth usage period only has an (assumed) start date and no end date, the duration of the period is assumed to be the median interval of the entire time series. If the user is unhappy with this assumption and they know the end date for the last period, they can add an additional row to their input data with the end date and 0 usage, then remove the last row of the output dataframe.