Overhaul to create_dataframe (switch to aggregate.r as the engine)
This update is an overhaul to the create_dataframe function. In an effort to simplify the aggregation process, remove dependence on the XTS package, and more accurately deal with time shifting, the majority of work has been handed to the recently revamped aggregate.r function. create_dataframe now serves as the front end function that provides error trapping and catching on user inputs (arguments). Aggregate.r is now the backend engine for this function. Additionally, a new argument has been added: shift_normal_weather. Rather than assuming the user always wants to shift the weather file from end of period reporting to start of period reporting, we now have an argument that defaults to false and can be specified as true to execute the shifting process. Shifting now happens prior to aggregation instead of after aggregation, addressing issue "Create dataframe shifts timestamps" opened on December 29 2022.