Small update to aggregate.R end date
Previously, if no end date was provided as an argument, it would be automatically generated based on the earliest ending date of all time series. This created an issue for longer interval data when aggregated to monthly intervals. For instance, if eload  has the earliest ending date at 5/15/23 and is in monthly intervals, then all other data including temp and additional independent variables would be cut to end at 5/15/23. This means for the final eload usage period (5/15/23-6/15/23) there would be only 1 day of temp and additional independent variable data to aggregate. This change addresses that issue by extending the end date by the eload interval when appropriate.