Simplifying logic in Aggregate.R for monthly aggregation
Previous development of monthly aggregation led to overly complicated logic and nested if blocks, each with their own aggregation procedures. This update significantly simplifies that structure by using the overlapping join method for all monthly data aggregation. This way, only a single set of aggregation actions have to be performed while if blocks handle the construction of intervals to aggregate to. This should hopefully improve readability and make the code more maintainable.

An important feature that comes from continued development of start and end date argument support is that now, if you use eload data that is originally in monthly+ increments, you can now specify when the final usage period ends by using the end_date argument. This will ensure that temperature and additional independent variables are correctly aggregated to the final usage period. If end_date is not provided, the function will attempt to determine when the final usage period ends based on the median interval of all preceeding usage periods.