Time zone error trapping for create_dataframe.R
This update organizes and expands error trapping for create_dataframe.R input arguments. The main addition is error trapping for mis-matched time zones. If dataframes are supplied to the function that have differing time zones, the function will throw an error and provide context with relevant time zones for each dataframe.